requirements.txt
appdirs==1.4.4
appnope==0.1.0
asgiref==3.2.7
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.8.2
black==19.10b0
certifi==2020.4.5.1
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
decorator==4.4.2
distlib==0.3.1
Django==3.0.6
django-debug-toolbar==2.2
django-extensions==2.2.9
django-modelcluster==5.0.1
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
filelock==3.0.12
flake8==3.8.3
html5lib==1.0.1
identify==1.4.21
idna==2.9
importlib-metadata==1.7.0
importlib-resources==3.0.0
ipython==7.15.0
ipython-genutils==0.2.0
isort==5.0.3
jedi==0.17.0
l18n==2018.5
mccabe==0.6.1
nodeenv==1.4.0
parso==0.7.0
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pre-commit==2.6.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.6.8
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.9
virtualenv==20.0.25
wagtail==2.9
wcwidth==0.2.4
webencodings==0.5.1
Willow==1.3
XlsxWriter==1.2.8
zipp==3.1.0
